// src/coaching/enums/coacheeErrors.enum.ts
export enum CoacheeErrors {
  COACHEE_FROM_ANOTHER_ORGANIZATION = 'COACHEE_FROM_ANOTHER_ORGANIZATION',
  COACHEE_PROFILE_SUSPENDED = 'COACHEE_PROFILE_SUSPENDED',
  COACHEE_NOT_ASSIGNED_TO_COACH = 'COACHEE_NOT_ASSIGNED_TO_COACH',
  NO_COACHEES_ASSIGNED = 'NO_COACHEES_ASSIGNED',
  NOT_COACHEE_EVALUATION_OWNER = 'NOT_COACHEE_EVALUATION_OWNER',
  NO_COACHEE_PROFILE = 'NO_COACHEE_PROFILE',
  NO_COACHEE_ROLE = 'NOT_COACHEE_ROLE',
  NOT_FOUND_COACHEE = 'NOT_FOUND_COACHEE',
  NOT_EXISTING_COACHEE = 'NOT_EXISTING_COACHEE',
  COACHEE_ALREADY_HAS_COACH = 'COACHEE_ALREADY_HAS_COACH',
  NOT_ASSIGNED_COCHEE = 'NOT_ASSIGNED_COCHEE',
  NOT_EXISTING_COCHEES = 'NOT_EXISTING_COCHEES',
  SUSPENDED_COACHEE = 'SUSPENDED_COACHEE',
  NOT_OWNER_ORGANIZATION_SUSPEND_COACHEE = 'NOT_OWNER_ORGANIZATION_SUSPEND_COACHEE',
  NOT_OWNER_ORGANIZATION_ACTIVATE_COACHEE = 'NOT_OWNER_ORGANIZATION_ACTIVATE_COACHEE',
  COACHEE_ALREADY_SUSPENDED = 'COACHEE_ALREADY_SUSPENDED',
  COACHEE_ALREADY_ACTIVE = 'COACHEE_ALREADY_ACTIVE',
  COACHEE_NOT_INVITED = 'COACHEE_NOT_INVITED',
  NOT_FOUND_COACHEE_TO_ACTIVATE = 'NOT_FOUND_COACHEE_TO_ACTIVATE',
  NOT_FOUND_COACHEE_TO_SUSPEND = 'NOT_FOUND_COACHEE_TO_SUSPEND',
  //
  NOT_OBJECTIVE_OWNER = 'NOT_OBJECTIVE_OWNER',
  //
  COACHEE_NOT_IN_COACHING_SESSION = 'COACHEE_NOT_IN_COACHING_SESSION',
  COACHEE_NOT_ADMIN = 'COACHEE_NOT_ADMIN',
}