// src/videoSessions/enums/coachingSessionFeedbackErrors.enum.ts
export enum CoachingSessionFeedbackErrors {
  COACHEE_NOT_IN_COACHING_SESSION = 'COACHEE_NOT_IN_COACHING_SESSION',
  COACH_NOT_IN_COACHING_SESSION = 'COACH_NOT_IN_COACHING_SESSION',
  NOT_ALL_QUESTIONS_WERE_ANSWERED = 'NOT_ALL_QUESTIONS_WERE_ANSWERED',
  UNEXECUTED_COACHING_SESSION = 'UNEXECUTED_COACHING_SESSION',
  ANSWER_DIFFER_FROM_QUESTIONS = 'ANSWER_DIFFER_FROM_QUESTIONS',
  SESSION_ALREADY_RATED = 'SESSION_ALREADY_RATED',
}